# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface DashboardSummary {
  greetingName: string;
  studyMinutesToday: number;
  tasksDueToday: number;
  upcomingExamsCount: number;
  focusStreakDays: number;
  bloomProgressPercent: number;
  bloomProgressLabel: string;
}

export type TaskType = typeof TaskType[keyof typeof TaskType];


export const TaskType = {
  homework: 'homework',
  note: 'note',
} as const;

export interface Task {
  id: number;
  title: string;
  type: TaskType;
  /** @nullable */
  subject?: string | null;
  /**
     * Calendar date YYYY-MM-DD
     * @nullable
     */
  dueDate?: string | null;
  completed: boolean;
  createdAt: string;
}

export type TaskInputType = typeof TaskInputType[keyof typeof TaskInputType];


export const TaskInputType = {
  homework: 'homework',
  note: 'note',
} as const;

export interface TaskInput {
  /** @minLength 1 */
  title: string;
  type: TaskInputType;
  subject?: string;
  /** Calendar date YYYY-MM-DD */
  dueDate?: string;
}

export interface TaskUpdate {
  /** @minLength 1 */
  title?: string;
  subject?: string;
  dueDate?: string;
  completed?: boolean;
}

export interface Exam {
  id: number;
  subject: string;
  /** Calendar date YYYY-MM-DD */
  examDate: string;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface ExamInput {
  /** @minLength 1 */
  subject: string;
  examDate: string;
  notes?: string;
}

export type ScheduleEventType = typeof ScheduleEventType[keyof typeof ScheduleEventType];


export const ScheduleEventType = {
  homework: 'homework',
  study_session: 'study_session',
  event: 'event',
  note: 'note',
} as const;

export interface ScheduleEvent {
  id: number;
  title: string;
  type: ScheduleEventType;
  /** Calendar date YYYY-MM-DD */
  date: string;
  /** 24h time HH:mm */
  time: string;
  createdAt: string;
}

export type ScheduleEventInputType = typeof ScheduleEventInputType[keyof typeof ScheduleEventInputType];


export const ScheduleEventInputType = {
  homework: 'homework',
  study_session: 'study_session',
  event: 'event',
  note: 'note',
} as const;

export interface ScheduleEventInput {
  /** @minLength 1 */
  title: string;
  type: ScheduleEventInputType;
  date: string;
  time: string;
}

export interface StudySession {
  id: number;
  /** @nullable */
  subject: string | null;
  durationMinutes: number;
  /** Calendar date YYYY-MM-DD */
  date: string;
  createdAt: string;
}

export interface StudySessionInput {
  subject?: string;
  /** @minimum 1 */
  durationMinutes: number;
  /** Calendar date YYYY-MM-DD, defaults to today */
  date?: string;
}

export interface Quote {
  text: string;
  /** @nullable */
  author?: string | null;
}

export type ListTasksParams = {
dueToday?: boolean;
};

export type ListExamsParams = {
upcoming?: boolean;
};

export type ListScheduleEventsParams = {
/**
 * Calendar date (YYYY-MM-DD). Defaults to today.
 */
date?: string;
};