# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardSummary {
  greetingName: string;
  studyMinutesToday: number;
  tasksDueToday: number;
  upcomingExamsCount: number;
  focusStreakDays: number;
  bloomProgressPercent: number;
  bloomProgressLabel: string;
}