# lib/api-zod/src/generated/types/exam.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Exam {
  id: number;
  subject: string;
  /** Calendar date YYYY-MM-DD */
  examDate: string;
  /** @nullable */
  notes?: string | null;
  createdAt: Date;
}