# lib/api-zod/src/generated/types/scheduleEvent.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ScheduleEventType } from './scheduleEventType';

export interface ScheduleEvent {
  id: number;
  title: string;
  type: ScheduleEventType;
  /** Calendar date YYYY-MM-DD */
  date: string;
  /** 24h time HH:mm */
  time: string;
  createdAt: Date;
}