# lib/api-zod/src/generated/types/scheduleEventInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ScheduleEventInputType } from './scheduleEventInputType';

export interface ScheduleEventInput {
  /** @minLength 1 */
  title: string;
  type: ScheduleEventInputType;
  date: string;
  time: string;
}