# lib/api-zod/src/generated/types/studySession.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface StudySession {
  id: number;
  /** @nullable */
  subject: string | null;
  durationMinutes: number;
  /** Calendar date YYYY-MM-DD */
  date: string;
  createdAt: Date;
}