# lib/api-zod/src/generated/types/taskInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TaskInputType } from './taskInputType';

export interface TaskInput {
  /** @minLength 1 */
  title: string;
  type: TaskInputType;
  subject?: string;
  /** Calendar date YYYY-MM-DD */
  dueDate?: string;
}